cc_requirements.txt
asgiref==3.2.7
codegen==1.0
Django==3.0.5
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-mysql==3.3.0
dnspython==1.16.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.2
patterns==0.3
Pillow==7.1.1
pkg-resources==0.0.0
protobuf==3.6.1
psutil==5.7.0
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1



phpmyadmin@localhost

mysqld is running as pid 19839

MySQL connection id 16
Server version: 8.0.20-0ubuntu0.20.03.1 (Ubuntu)

$lsb_release -a
No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 20.04 LTS
Release:        20.04
Codename:       focal 